# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Baliarda Blister Defect Tracking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DefectType = (typeof DefectType)[keyof typeof DefectType];

export const DefectType = {
  arrugas: "arrugas",
  pisados: "pisados",
  codificado_cortado: "codificado_cortado",
  codificado_poco_legible: "codificado_poco_legible",
  blisters_vacio: "blisters_vacio",
  blisters_ausencia_comprimido: "blisters_ausencia_comprimido",
  blisters_comprimido_partido: "blisters_comprimido_partido",
  poco_segrinado: "poco_segrinado",
  polvo: "polvo",
  manchas: "manchas",
  pinchados: "pinchados",
  comprimido_con_pelo: "comprimido_con_pelo",
} as const;

export interface DefectItem {
  type: string;
  count: number;
}

export interface Defect {
  id: number;
  opNumber: string;
  bulkCode: string;
  product: string;
  lot: string;
  orderQuantity: number;
  blistersPerBox: number;
  totalBlisters: number;
  defectiveBlisters: number;
  incidenceRate: number;
  defectType: string;
  defectItems: DefectItem[];
  labelPhotoUrl?: string | null;
  defectPhotoUrls: string[];
  observations?: string | null;
  createdAt: string;
}

export interface CreateDefectBody {
  opNumber: string;
  bulkCode: string;
  product: string;
  lot: string;
  orderQuantity: number;
  blistersPerBox: number;
  totalBlisters: number;
  defectiveBlisters: number;
  incidenceRate: number;
  defectType: string;
  defectItems: DefectItem[];
  labelPhotoUrl?: string | null;
  defectPhotoUrls: string[];
  observations?: string | null;
}

export interface UpdateDefectBody {
  opNumber: string;
  bulkCode: string;
  product: string;
  lot: string;
  orderQuantity: number;
  blistersPerBox: number;
  totalBlisters: number;
  defectiveBlisters: number;
  incidenceRate: number;
  defectType: string;
  defectItems: DefectItem[];
  labelPhotoUrl?: string | null;
  defectPhotoUrls: string[];
  observations?: string | null;
}

export type DefectStatsByTypeItem = {
  defectType: string;
  count: number;
  totalDefective: number;
};

export type DefectStatsByDateItem = {
  date: string;
  count: number;
  totalDefective: number;
};

export interface DefectStats {
  totalRecords: number;
  totalBlistersInspected: number;
  totalDefective: number;
  averageIncidenceRate: number;
  byType: DefectStatsByTypeItem[];
  byDate: DefectStatsByDateItem[];
}

export interface UploadPhotoBody {
  /** Base64 encoded image data */
  data: string;
  mimeType: string;
}

export interface PhotoUploadResult {
  url: string;
}

export type ListDefectsParams = {
  fromDate?: string;
  toDate?: string;
  defectType?: string;
};

export type GetDefectStatsParams = {
  fromDate?: string;
  toDate?: string;
};

export type ExportDefectsExcelParams = {
  fromDate?: string;
  toDate?: string;
  defectType?: string;
};