# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Baliarda Blister Defect Tracking
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all defect records
 */
export const ListDefectsQueryParams = zod.object({
  fromDate: zod.date().optional(),
  toDate: zod.date().optional(),
  defectType: zod.coerce.string().optional(),
});

export const ListDefectsResponseItem = zod.object({
  id: zod.number(),
  opNumber: zod.string(),
  bulkCode: zod.string(),
  product: zod.string(),
  lot: zod.string(),
  orderQuantity: zod.number(),
  blistersPerBox: zod.number(),
  totalBlisters: zod.number(),
  defectiveBlisters: zod.number(),
  incidenceRate: zod.number(),
  defectType: zod.string(),
  defectItems: zod.array(
    zod.object({
      type: zod.string(),
      count: zod.number(),
    }),
  ),
  labelPhotoUrl: zod.string().nullish(),
  defectPhotoUrls: zod.array(zod.string()),
  observations: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListDefectsResponse = zod.array(ListDefectsResponseItem);

/**
 * @summary Create a new defect record
 */
export const CreateDefectBody = zod.object({
  opNumber: zod.string(),
  bulkCode: zod.string(),
  product: zod.string(),
  lot: zod.string(),
  orderQuantity: zod.number(),
  blistersPerBox: zod.number(),
  totalBlisters: zod.number(),
  defectiveBlisters: zod.number(),
  incidenceRate: zod.number(),
  defectType: zod.string(),
  defectItems: zod.array(
    zod.object({
      type: zod.string(),
      count: zod.number(),
    }),
  ),
  labelPhotoUrl: zod.string().nullish(),
  defectPhotoUrls: zod.array(zod.string()),
  observations: zod.string().nullish(),
});

/**
 * @summary Update an existing defect record
 */
export const UpdateDefectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDefectBody = zod.object({
  opNumber: zod.string(),
  bulkCode: zod.string(),
  product: zod.string(),
  lot: zod.string(),
  orderQuantity: zod.number(),
  blistersPerBox: zod.number(),
  totalBlisters: zod.number(),
  defectiveBlisters: zod.number(),
  incidenceRate: zod.number(),
  defectType: zod.string(),
  defectItems: zod.array(
    zod.object({
      type: zod.string(),
      count: zod.number(),
    }),
  ),
  labelPhotoUrl: zod.string().nullish(),
  defectPhotoUrls: zod.array(zod.string()),
  observations: zod.string().nullish(),
});

export const UpdateDefectResponse = zod.object({
  id: zod.number(),
  opNumber: zod.string(),
  bulkCode: zod.string(),
  product: zod.string(),
  lot: zod.string(),
  orderQuantity: zod.number(),
  blistersPerBox: zod.number(),
  totalBlisters: zod.number(),
  defectiveBlisters: zod.number(),
  incidenceRate: zod.number(),
  defectType: zod.string(),
  defectItems: zod.array(
    zod.object({
      type: zod.string(),
      count: zod.number(),
    }),
  ),
  labelPhotoUrl: zod.string().nullish(),
  defectPhotoUrls: zod.array(zod.string()),
  observations: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get a single defect record
 */
export const GetDefectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDefectResponse = zod.object({
  id: zod.number(),
  opNumber: zod.string(),
  bulkCode: zod.string(),
  product: zod.string(),
  lot: zod.string(),
  orderQuantity: zod.number(),
  blistersPerBox: zod.number(),
  totalBlisters: zod.number(),
  defectiveBlisters: zod.number(),
  incidenceRate: zod.number(),
  defectType: zod.string(),
  defectItems: zod.array(
    zod.object({
      type: zod.string(),
      count: zod.number(),
    }),
  ),
  labelPhotoUrl: zod.string().nullish(),
  defectPhotoUrls: zod.array(zod.string()),
  observations: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a defect record
 */
export const DeleteDefectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get defect statistics summary
 */
export const GetDefectStatsQueryParams = zod.object({
  fromDate: zod.date().optional(),
  toDate: zod.date().optional(),
});

export const GetDefectStatsResponse = zod.object({
  totalRecords: zod.number(),
  totalBlistersInspected: zod.number(),
  totalDefective: zod.number(),
  averageIncidenceRate: zod.number(),
  byType: zod.array(
    zod.object({
      defectType: zod.string(),
      count: zod.number(),
      totalDefective: zod.number(),
    }),
  ),
  byDate: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
      totalDefective: zod.number(),
    }),
  ),
});

/**
 * @summary Export defects to Excel
 */
export const ExportDefectsExcelQueryParams = zod.object({
  fromDate: zod.date().optional(),
  toDate: zod.date().optional(),
  defectType: zod.coerce.string().optional(),
});

/**
 * @summary Upload a photo (base64)
 */
export const UploadPhotoBody = zod.object({
  data: zod.string().describe("Base64 encoded image data"),
  mimeType: zod.string(),
});